Trying to debug USB4000 bug when opening multiple spectrometers.
git-svn-id: https://cotprojects.marine.usf.edu/svn/MiniSEAS/Gumstix/Source@26 63cbca5c-2ead-4e1b-9715-33247cc11c71

+++ USB4000Gum/source/USB4000Gum.h
@@ -55,7 +55,7 @@ typedef struct CALIBRATIONCOEFFICIENTS{
 
 typedef struct SPECSSAMPLE{
     unsigned short numScansForSample;
-    short* pixels;
+    float* pixels;
 } specSample;
 
 typedef struct SPECTROMETER{
